+++ assembly/utils/types/DataConnectorConfig.ts
@@ -1,0 +1,7 @@
+import { JSON } from 'json-as/assembly';
+import { ExecutionContext } from "./ExecutionContext";
+
+@serializable
+export class DataConnectorConfig {
+  executionContext: ExecutionContext = new ExecutionContext();
+}