+++ assembly/utils/types/index.ts
@@ -1,2 +1,5 @@
+export * from './Candle';
+export * from './ExecutionContext';
+export * from './DataConnectorConfig';
 export * from './Position';
-export * from './Price';
+export * from './RawTradeData';